Massive commit… make oc_heimdall work
This commit adds support for actually building the oc_heimdall release
(now that we have Erlang code written).

Configuration information is managed by a template, as are the various
scripts needed to actually run the software (as opposed to using
scripts embedded in the project repo itself).

Additionally, templates and recipes are tailored in such a way that
they could be extracted into a future 'opscode_erlang_app' cookbook,
which would define the "Opscode Way" of deploying an Erlang
application.