Break up release build steps
When we're in development mode (and want quicker turn-around on
provisions), we'll just call 'make relclean rel' to rebuild the
release.  When in a non-development mode, we'll preface that with a
'make distclean', in order to ensure that we always have the proper
dependencies in place.

If you're developing and add a new dependency, you'll either have to
run 'rebar get-deps' manually, or just rebuild the VM from scratch.
This will be a relatively rare thing to do, however, and the speedup
for the more common development case is worth it.