Updates for latest in Test Kitchen 1.0
Strictly speaking, test-kitchen does not need to be used w/ bundler.
This has the side effect of "just working" in any updated cookbook
repository as long as the gems needed already are installed. It also
means that the vagrant as a gem and vagrantfile are not required.

* No longer use vagrant as a gem (Gemfile, Vagrantfile)
* Point to the prerelease gem version of test-kitchen
* Point to the released version of kitchen-vagrant
* Ensure omnibus package is installed in the vagrant boxes
* Use opsmaster config for chef_api

You'll need to copy your opsmaster user key to
~/.chef/$USER-opsmaster.pem, where $USER is your username which
*should* be your opsmaster-api OPC username.

The default suite is now named deploy, though the data bag doesn't
load so kitchen fails to converge.