Relax chef-vault version to minor , not patch
In 620b1a32, chef-vault was added for AWS secrets. The version
constraint was used to ensure that the `dev_mode` fallback capability
of the chef-vault cookbook was available.

The pessimistic operator against the patch version is too restrictive.
We should be able to use a cookbook that we maintain (chef-vault) of
later minor versions without issue, but we will want to avoid major
version updates.